+++ src/containers/PlayerContainer.jsx
@@ -1,8 +1,8 @@
 import React, {PureComponent} from 'react';
 import {connect} from 'react-redux';
-import {dom, react, redux} from '../helpers/helpers';
-import collision from '../helpers/collision';
-import offset from '../helpers/offset';
+import {_react_, _redux_} from '../helpers/helpers';
+import _collision_ from '../helpers/collision';
+import _offset_ from '../helpers/offset';
 import PlayerView from '../views/PlayerView';
 import LightRadiusView from '../views/LightRadiusView';
 
@@ -10,18 +10,13 @@ class PlayerContainer extends PureComponent {
 
   /* lifecycle */
 
-  constructor(props) {
-    super(props);
-    this.roomComponent = null;
-    this.state = {
-      playerProps: null,
-      lightRadiusProps: null
-    }
-  }
+  roomComponent = null;
+  state = {
+    playerProps: null,
+    lightRadiusProps: null
+  };
 
   componentDidMount() {
-    this.addKeyboardSupport();
-
     setTimeout(() => {
       this.props.actions.Tile.selectTile(0, 0);
       this.props.actions.Player.updateLightRadius(4);
@@ -29,55 +24,62 @@ class PlayerContainer extends PureComponent {
   }
 
   componentWillReceiveProps(nextProps) {
-    this.setPlayerPosition(nextProps);
-    this.setTileCollision(nextProps);
-    this.setPlayerOffset(nextProps);
+    const _update_ = {props: this.props, contextName: this.constructor.name};
+    this.updatePlayerPosition({...nextProps, ...{ ..._update_, methodName: this.updatePlayerPosition.name }});
+    this.updateTileCollision({...nextProps, ...{ ..._update_, methodName: this.updateTileCollision.name }});
+    this.updatePlayerOffset({...nextProps, ...{ ..._update_, methodName: this.updatePlayerOffset.name }});
   }
 
-  /* local state */
+  updatePlayerPosition(nextProps) {
+    if (_react_.stateDidUpdate(nextProps, 'Tile.tileId') ||
+      _react_.stateDidUpdate(nextProps, 'Tile.roomId')) {
+      this.props.actions.Player.updatePosition();
+    }
+  }
 
-  getPlayerProps(style, lightRadiusStyle) {
-    this.setState({
-      ...this.state,
-      playerProps: {
-        style,
-        id: 0
-      },
-      lightRadiusProps: {
-        style: lightRadiusStyle
+  updateTileCollision(nextProps) {
+    if (_react_.stateDidUpdate(nextProps, 'GameBoard.level')) {
+      // reset and recalculate collisions on level change
+      this.roomComponent = _collision_.resetCollision(this.roomComponent);
+      const tileId = _.get(nextProps, 'state.Entity.Player.spawns[0].position.tileId');
+      if (tileId === 0) {
+        this.calculateTileCollision(nextProps);
+      } else if (tileId !== undefined) {
+        this.props.actions.Tile.selectTile(0, 0);
       }
-    });
+    } else if (_react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].position.roomId')) {
+      // reset and recalculate collisions on room change
+      this.roomComponent = _collision_.resetCollision(this.roomComponent);
+      this.calculateTileCollision(nextProps);
+    } else if (_react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].position.tileId') ||
+      _react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].lightRadius')) {
+      // recalculate collisions on tile or radius change
+      this.calculateTileCollision(nextProps);
+    }
+  }
+
+  updatePlayerOffset(nextProps) {
+    if (_react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].position.tileId') ||
+      _react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].position.roomId') ||
+      _react_.stateDidUpdate(nextProps, 'Entity.Player.spawns[0].lightRadius') ||
+      _react_.stateDidUpdate(nextProps, 'GameBoard.hasPerspective') ||
+      _react_.stateDidUpdate(nextProps, 'GameBoard.level')) {
+      this.calculatePlayerOffset(nextProps);
+    }
   }
 
   /* methods */
 
-  addKeyboardSupport() {
-    window.addEventListener('keydown', e => {
-      if (e.defaultPrevented) return;
-      e.preventDefault();
-
-      const index = collision.getTileIndex(
-        this.roomComponent,
-        this.props.state.Entity.Player.spawns[0].position.tileId,
-        e.key
-      );
-      if (!isNaN(index)) this.props.actions.Tile.selectTile(
-        index,
-        this.props.state.Entity.Player.spawns[0].position.roomId
-      );
-    }, true);
-  }
-
   getRoomComponent(props) {
     const roomId = props.state.Entity.Player.spawns[0].position.roomId;
     const roomElm = document.querySelector(`#room${roomId}`);
-    if (!isNaN(roomId) && roomElm) this.roomComponent = dom.getComponent(roomElm);
+    if (!isNaN(roomId) && roomElm) this.roomComponent = _react_.getComponent(roomElm);
   }
 
   calculateTileCollision(props) {
     if (!this.roomComponent) this.getRoomComponent(props);
 
-    collision.tileCollision(
+    _collision_.tileCollision(
       this.roomComponent,
       props.state.Entity.Player.spawns[0].position.tileId,
       props.state.Entity.Player.spawns[0].lightRadius,
@@ -90,54 +92,24 @@ class PlayerContainer extends PureComponent {
 
     const tileId = props.state.Entity.Player.spawns[0].position.tileId;
     if (!isNaN(tileId)) {
-      const style = offset.playerOffset(this.roomComponent, props);
-      const lightRadiusStyle = offset.lightRadiusOffset(this.roomComponent, props);
-      this.getPlayerProps(style, lightRadiusStyle);
+      const style = _offset_.playerOffset(this.roomComponent, props);
+      const lightRadiusStyle = _offset_.lightRadiusOffset(this.roomComponent, props);
+      this.setPlayerProps(style, lightRadiusStyle);
     }
   }
 
-  /* updates */
+  /* local state */
 
-  setPlayerPosition(nextProps) {
-    if (
-      react.stateDidUpdate(this.props, nextProps, 'Tile.tileId') ||
-      react.stateDidUpdate(this.props, nextProps, 'Tile.roomId')
-    ) {
-      this.props.actions.Player.updatePosition();
-    }
-  }
-
-  setTileCollision(nextProps) {
-    if (react.stateDidUpdate(this.props, nextProps, 'GameBoard.level')) {
-      // reset and recalculate collisions on level change
-      this.roomComponent = collision.resetCollision(this.roomComponent);
-      if (
-        nextProps.state.Entity.Player.spawns[0].position.tileId === 0
-      ) {
-        this.calculateTileCollision(nextProps);
-      } else {
-        this.props.actions.Tile.selectTile(0, 0);
+  setPlayerProps(style, lightRadiusStyle) {
+    this.setState({
+      ...this.state,
+      playerProps: {
+        style, id: 0
+      },
+      lightRadiusProps: {
+        style: lightRadiusStyle
       }
-    } else if (react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].position.roomId')) {
-      // reset and recalculate collisions on room change
-      this.roomComponent = collision.resetCollision(this.roomComponent);
-      this.calculateTileCollision(nextProps);
-    } else if (react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].position.tileId') ||
-      react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].lightRadius')) {
-      // recalculate collisions on tile or radius change
-      this.calculateTileCollision(nextProps);
-    }
-  }
-
-  setPlayerOffset(nextProps) {
-    if (
-      react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].position.tileId') ||
-      react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].position.roomId') ||
-      react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].lightRadius') ||
-      react.stateDidUpdate(this.props, nextProps, 'GameBoard.level')
-    ) {
-      this.calculatePlayerOffset(nextProps);
-    }
+    });
   }
 
   /* render */
@@ -146,17 +118,17 @@ class PlayerContainer extends PureComponent {
     return (
       <div className="player">
         <PlayerView {...this.state.playerProps} />
-        <LightRadiusView {...this.state.lightRadiusProps} />
+        {/*<LightRadiusView {...this.state.lightRadiusProps} />*/}
       </div>
     );
   }
 }
 
 export default connect(
-  state => redux.mapState(state, [
+  state => _redux_.mapState(state, [
     'Entity', 'Tile', 'GameBoard',
   ]),
-  dispatch => redux.mapActions(dispatch, [
+  dispatch => _redux_.mapActions(dispatch, [
     'Entity', 'Player', 'Tile'
   ])
 )(PlayerContainer);