+++ src/helpers/collision.js
@@ -1,31 +1,32 @@
-import collision from './collision';
+import _collision_ from './collision';
 
 export default {
 
   roomCollision: (room, rooms, ignored) => {
-    for (var i = 0; i < rooms.length; i++) {
+    for (let i = 0; i < rooms.length; i++) {
       if (i == ignored) continue;
       const check = rooms[i];
-      if (!((room.x + room.w < check.x) || (room.x > check.x + check.w) || (room.y + room.h < check.y) || (room.y > check.y + check.h))) return true;
+      if (!((room.x + room.w < check.x) ||
+          (room.x > check.x + check.w) ||
+          (room.y + room.h < check.y) ||
+          (room.y > check.y + check.h))
+      ) return true;
     }
     return false;
   },
 
   roomSquash: (rooms, tileSize) => {
-    for (var i = 0; i < 10; i++) {
-      for (var j = 0; j < rooms.length; j++) {
-        var room = rooms[j];
+    for (let i = 0; i < 10; i++) {
+      for (let j = 0; j < rooms.length; j++) {
+        const room = rooms[j];
         while (true) {
-          var old_position = {
-            x: room.x,
-            y: room.y
-          };
+          const oldPos = { x: room.x, y: room.y };
           if (room.x > 1) room.x -= tileSize;
           if (room.y > 1) room.y -= tileSize;
           if ((room.x < tileSize) && (room.y < tileSize)) break;
-          if (collision.roomCollision(room, rooms, j)) {
-            room.x = old_position.x;
-            room.y = old_position.y;
+          if (_collision_.roomCollision(room, rooms, j)) {
+            room.x = oldPos.x;
+            room.y = oldPos.y;
             break;
           }
         }
@@ -36,7 +37,8 @@ export default {
   tileCollision: (component, tileId, radius, tileSize) => {
     if (!component || isNaN(tileId) || isNaN(radius) || isNaN(tileSize)) return;
     // flatten the room into tiles
-    const tiles = component.instance.props.children ? Array.prototype.concat.apply([], component.instance.props.children) : null;
+    const tiles = component.instance.props.children
+      ? Array.prototype.concat.apply([], component.instance.props.children) : null;
 
     // get the tile where the light is centered
     const offsetIndex = tiles[0].props.children.props.id;
@@ -70,7 +72,7 @@ export default {
       const tileElm = component.elm.querySelector(`#tile${i+offsetIndex}`);
       if (ratio) {
         tileElm.setAttribute('data-light', 'on');
-        tileElm.style['opacity'] = (ratio > .62) ? 1 : (ratio < .2) ? .2 : ratio;
+        tileElm.style['opacity'] = (ratio > .72) ? 1 : (ratio < .2) ? .2 : ratio;
       } else if (_.get(tileElm, 'attributes[\'data-light\'].nodeValue') === 'on') {
         tileElm.setAttribute('data-light', 'off');
       }
@@ -87,32 +89,15 @@ export default {
     return null;
   },
 
-  getTileIndex: (component, tileId, pressedKey) => {
-    if (!component || isNaN(tileId) || !pressedKey) return;
+  getTileIndex: (component, tileId, delta) => {
+    if (!component || isNaN(tileId) || !delta) return;
     // flatten the room into tiles
     const tiles = component.instance.props.children ? Array.prototype.concat.apply([], component.instance.props.children) : null;
     const offsetIndex = tiles[0].props.children.props.id;
 
     const normalized = tileId - offsetIndex;
     const currentTile = tiles[normalized].props.children.props;
-    const coord = { x: currentTile.x, y: currentTile.y };
-
-    switch (pressedKey) {
-      case "ArrowDown":
-        coord.y += 1;
-        break;
-      case "ArrowUp":
-        coord.y -= 1;
-        break;
-      case "ArrowLeft":
-        coord.x -= 1;
-        break;
-      case "ArrowRight":
-        coord.x += 1;
-        break;
-      default:
-        return; // Quit when this doesn't handle the key event.
-    }
+    const coord = { x: currentTile.x + delta.x, y: currentTile.y + delta.y };
 
     const selectedTile = component.elm.querySelector(`[data-coord="${coord.x}-${coord.y}"]`);
     return selectedTile ? Number(selectedTile.id.replace('tile', '')) : undefined;