+++ src/views/GameBoardView.jsx
@@ -2,5 +2,5 @@ import React from 'react';
 import './GameBoardView.scss';
 
 export default props => (
-  <div className='game-board'>{ props.children }</div>
+  <div id='gameboard'>{ props.children }</div>
 );